% Generated by roxygen2: do not edit by hand
% Please edit documentation in logic/utilites.R
\name{create_dropdown_list}
\alias{create_dropdown_list}
\title{Create dropdown list}
\usage{
create_dropdown_list(df, distinct_col)
}
\arguments{
\item{df}{A dataframe containing the column to create dropdown list from}

\item{distinct_col}{The column to pull list from}
}
\description{
Creates a vector of unique values to be used in a shiny select input.
}
